feat: add unit test class for Util_Functions.py

+++ projects/API Based Weather Report/Util_Functions.py
@@ -60,7 +60,7 @@ def convert_temperature(str_temperature_kelvin, temperature_unit):
     """
     # Convert strings to integers
     try:
-        temperature_k = int(str_temperature_kelvin)
+        temperature_k = float(str_temperature_kelvin)
     except ValueError:
         return "API temperature data format error!"
